added singleEvent mode, restructure runfit.cpp
