+++ src/PMT_FindAlpha.cpp
@@ -11,8 +11,11 @@
 #include <BAT/BCMath.h>
 #include <cmath>
 
+#include "PMT_association.hpp"
+#include "helper_lib.hpp"
 
-// PMTfit class
+
+// PMTfindalpha class
 PMTfindalpha::PMTfindalpha(const std::string& mode, int nth, int nP, 
                 const std::vector<double>& L1_inp, const std::vector<double>& L2_inp, const std::vector<double>& L3_inp, 
                 const std::vector<double>& L4_inp, const std::vector<double>&x, const std::vector<double>& y): BCModel(mode)
@@ -81,18 +84,116 @@ double PMTfindalpha::LogLikelihood(const std::vector<double>& pars) {
 }
 
 
-// Function to calculate distance between the PMT and the chosen position
-double PMTfindalpha::D2(double x, double y, int i) {
-    if (i == 0) {
-        return (x - x1)*(x - x1) + (y - y1)*(y - y1) + zGEM*zGEM;
-    } else if (i == 1) {
-        return (x - x2)*(x - x2) + (y - y2)*(y - y2) + zGEM*zGEM;
-    } else if (i == 2) {
-        return (x - x3)*(x - x3) + (y - y3)*(y - y3) + zGEM*zGEM;
-    } else if (i == 3) {
-        return (x - x4)*(x - x4) + (y - y4)*(y - y4) + zGEM*zGEM;
-    } else {
-        throw std::runtime_error("Uknown value of PMT index.\n");
+// // Function to calculate distance between the PMT and the chosen position
+// double PMTfindalpha::D2(double x, double y, int i) {
+//     if (i == 0) {
+//         return (x - x1)*(x - x1) + (y - y1)*(y - y1) + zGEM*zGEM;
+//     } else if (i == 1) {
+//         return (x - x2)*(x - x2) + (y - y2)*(y - y2) + zGEM*zGEM;
+//     } else if (i == 2) {
+//         return (x - x3)*(x - x3) + (y - y3)*(y - y3) + zGEM*zGEM;
+//     } else if (i == 3) {
+//         return (x - x4)*(x - x4) + (y - y4)*(y - y4) + zGEM*zGEM;
+//     } else {
+//         throw std::runtime_error("Uknown value of PMT index.\n");
+//     }
+//     return 0.;
+// }
+
+
+void PMTfindalpha::runAlphaFit(const Config config){
+    // Setting chains parameters
+    int Nch = 12;          //number of parallel MCMC chains
+    int NIter = 100000;    //number of step per chain
+
+    // prepare helper variables
+    std::vector<double> L1;
+    std::vector<double> L2;
+    std::vector<double> L3;
+    std::vector<double> L4;
+    std::vector<double> x;
+    std::vector<double> y;
+
+    std::string mode = config.mode;
+    std::string input_file = config.input_file;
+    int start_ind = config.start_ind;
+    int end_ind = config.end_ind;
+    int nPoints = config.nPoints;
+    std::string output_tag = config.calibration_output_tag;
+
+    //  Create results folder if not existing
+    std::string res_dir;
+    res_dir = "./output_"+mode;
+    int com = std::system(("mkdir -p "+res_dir).c_str());
+    if(com == 0) {
+    std::cerr << "Failed to create directory: " << res_dir << std::endl;
     }
-    return 0.;
+    res_dir += "/";
+
+    // loop over the nPoints points to store data
+    for(int point = 0; point<end_ind; point++){
+        // HERE I NORMALIZE THE Li USING SC_INTEGRAL, SO THAT IF THE LY IS NOT CONSTANT IS ALL GOOD
+        L1.push_back(data.getL1()[point]/data.getSc_integral()[point]*10000);
+        L2.push_back(data.getL2()[point]/data.getSc_integral()[point]*10000);
+        L3.push_back(data.getL3()[point]/data.getSc_integral()[point]*10000);
+        L4.push_back(data.getL4()[point]/data.getSc_integral()[point]*10000);
+        
+        x.push_back(data.getXtrue()[point]);
+        y.push_back(data.getYtrue()[point]);
+    }
+
+    // Create log
+    BCLog::OpenLog("./logs/findalpha_log.txt", BCLog::detail, BCLog::detail);
+
+    // INITIALIZE THE MODEL
+    PMTfindalpha alpha(mode, Nch, end_ind, L1, L2, L3, L4, x, y);
+
+    // Setting MCMC algorithm and precision
+    alpha.SetMarginalizationMethod(BCIntegrate::kMargMetropolis);
+    alpha.SetPrecision(BCEngineMCMC::kMedium);
+
+    BCLog::OutSummary("Model created");
+
+    // Setting prerun iterations to 10^6
+    alpha.SetNIterationsPreRunMax(1000000);
+    
+    // Setting MC run iterations and number of parallel chains
+    alpha.SetNIterationsRun(NIter);
+    alpha.SetNChains(Nch);
+    
+    // Prefix for BAT outputs
+    std::string BAT_out_prefix_alpha = res_dir+alpha.GetSafeName() + "_" + output_tag;
+
+    alpha.WriteMarkovChain(BAT_out_prefix_alpha + "_mcmc.root", "RECREATE");
+
+    // ===============================================================
+    // Run MCMC, marginalizing posterior
+    alpha.MarginalizeAll();
+    // ===============================================================
+
+    // Run mode finding; by default using Minuit
+    alpha.FindMode(alpha.GetBestFitParameters());
+    
+    if (plot) {
+        // Draw all marginalized distributions into a PDF file
+        alpha.PrintAllMarginalized(BAT_out_prefix_alpha + "_plots.pdf");
+            
+        // Print summary plots
+        alpha.PrintParameterPlot(BAT_out_prefix_alpha + "_parameters.pdf");
+        alpha.PrintCorrelationPlot(BAT_out_prefix_alpha + "_correlation.pdf");
+        alpha.PrintCorrelationMatrix(BAT_out_prefix_alpha + "_correlationMatrix.pdf");
+        alpha.PrintKnowledgeUpdatePlots(BAT_out_prefix_alpha + "_update.pdf");
+    }
+    
+    // Print results of the analysis
+    alpha.PrintSummary();
+    
+    // Close log file
+    BCLog::OutSummary("Exiting");
+    BCLog::CloseLog();
+    
+    return 0;
+        
+
 }
+
